Pomodoro에서 시간 설정을 Picker로 처리되도록 수정

+++ HabitTimer/Common/Config.swift
@@ -22,7 +22,7 @@ class Config {
     
     static let TODOLIST_ROW_HEIGHT: CGFloat = 340
     
-    static let POMODORO_WORK_TIME_MINUTE: Double = 2 //25
+    static let POMODORO_WORK_TIME_MINUTE: Int = 2 //25
     
     static let POMODORO_REST_TIME_MINUTE: Double = 5
